requirements.txt
Flask==0.12.2
flask_wtf==0.14.3
PILLOW==7.2.0
numpy==1.18.5
matplotlib==3.3.1
gunicorn==20.0.4
tensorflow==2.3.0
Keras==2.4.3
